+++ exec/java-exec/src/main/resources/drill-module.conf
@@ -75,6 +75,7 @@ drill.exec: {
     implementation: "org.apache.drill.exec.opt.IdentityOptimizer"
   },
   storage: {
+    registry: "org.apache.drill.exec.store.StoragePluginRegistryImpl",
     file: {
       text: {
         buffer.size: 262144,
